move resolvers into gqlgen package
